% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as_inequalities}
\alias{as_inequalities}
\title{transform Hasse diagram to paritial order inequalities}
\usage{
as_inequalities(Hasse, Z)
}
\arguments{
\item{Hasse}{the partial order relations expressed by directed graph}

\item{Z}{a matrix of staging variable (poset)}
}
\value{
a matrix expressing the patial order constraints inequalities on coefficients
}
\description{
transform Hasse diagram to paritial order inequalities
}
