% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bestConvex}
\alias{bestConvex}
\title{best convex oracle}
\usage{
bestConvex(y, experts, awake = NULL, loss.type = "squareloss", niter = 1,
  method = "L-BFGS-B", control = list(maxit = 100, trace = T))
}
\arguments{
\item{y}{vector that contains the observations
to be predicted.}

\item{experts}{A matrix containing the
experts forecasts. Each column corresponds to the predictions proposed by an
expert to predict \code{Y}. It has as many columns as there are experts.}

\item{awake}{A matrix specifying the
activation coefficients of the experts. Its entries lie in \code{[0,1]}.
Needed if some experts are specialists and do not always form and suggest
prediction.  If the expert number \code{k} at instance \code{t} does not
form any prediction of observation \code{Y_t}, we can put
\code{awake[t,k]=0} so that the mixture does not consider expert \code{k} in
the mixture to predict \code{Y_t}.}

\item{loss.type}{A string specifying
the loss function considered to evaluate the performance. It can be
"squareloss", "mae", "mape", or "pinballloss". See \code{\link{loss}} for
more details.}

\item{niter}{Number of times the convex
optimisation process is repeated with different initial values. Note that if
the experts are always activated, it does not need to be greater than 1.}

\item{method}{The optimization method to be used. See \code{\link{optim}}.}

\item{control}{Additional parameters
that are passed to \code{optim} function.}
}
\value{
\item{loss}{ The loss suffered by the best fixed convex combination.
} \item{weights}{ A vector containing the best fixed convex combination
chosen in hindsight.  } \item{prediction}{ A vecor containing the
predictions of the best fixed convex combination.  }
}
\description{
The
function \code{bestConvex} computes the best convex combination oracle, or
in other words the fixed convex combination of experts in \code{experts}
that would have perfomed the best to predict the observations in \code{y}.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{bestLinear}}, \code{\link{bestShifts}}, \code{\link{loss}},
\code{\link{lossConv}}
}
\keyword{~kwd1}
\keyword{~kwd2}

