% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_matrix.R
\name{check_matrix}
\alias{check_matrix}
\title{Function to check and modify the input class and type}
\usage{
check_matrix(mat, name)
}
\arguments{
\item{mat}{\code{data.frame, data.table, tibble}. Object to be cast to matrix.}

\item{name}{\code{character}. Name of the object to be cast.}
}
\value{
a 3d array if a 3d array is provided, else a matrix.
}
\description{
Function to check and modify the input class and type
}
