% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convexLoss}
\alias{convexLoss}
\title{agerage error suffered by convex combination}
\usage{
convexLoss(weights, y, experts, awake = NULL)
}
\arguments{
\item{weights}{vector of weights
assigned to each experts. It must be the same length as the number of
columns of \code{experts}. It entries must be non-negative and sum to one.}

\item{y}{A vector containing the observations
to be predicted.}

\item{experts}{A matrix containing the
experts forecasts. Each column corresponds to the predictions proposed by an
expert to predict \code{Y}. It has as many columns as there are experts.}

\item{awake}{A matrix specifying the
activation coefficients of the experts. Its entries lie in \code{[0,1]}.
Needed if some experts are specialists and do not always form and suggest
prediction.  If the expert number \code{k} at instance \code{t} does not
form any prediction of observation \code{Y_t}, we can put
\code{awake[t,k]=0} so that the mixture does not consider expert \code{k} in
the mixture to predict \code{Y_t}.}
}
\value{
A dataframe containing several measures of losses suffered
by the mixture
}
\description{
The
function \code{convexLoss} computes the loss incured by a fixed convex
combination of experts.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{loss}}
}
\keyword{~kwd1}
\keyword{~kwd2}

