% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cumsum_rev}
\alias{cumsum_rev}
\title{calculate reversed cumulative sum}
\usage{
cumsum_rev(a)
}
\arguments{
\item{a}{vector}
}
\value{
a vector of reversed cumulative sum
}
\description{
calculate reversed cumulative sum
}
\examples{
cumsum_rev(1:10)
55 54 52 49 45 40 34 27 19 10


}
