% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggopera.R
\name{ggopera.mixture}
\alias{ggopera.mixture}
\title{Plot an object of class mixture with ggplot2}
\usage{
\method{ggopera}{mixture}(x, type = 1, dates = NULL, col = NULL)
}
\arguments{
\item{x}{an object of class mixture. If awake is provided (i.e., some experts are unactive), 
their residuals and cumulative losses are computed by using the predictions of the mixture.}

\item{type}{which graphic to plot, a value between 1 and 6.}

\item{dates}{a vector of dates to represent the x axis labels,}

\item{col}{the color to use to represent each experts, if set to NULL (default) use R\code{RColorBrewer::brewer.pal(...,"Spectral"}}
}
\value{
A ggplot object
}
\description{
provides different diagnostic plots for an aggregation procedure.
}
