% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggopera.R
\name{ggopera}
\alias{ggopera}
\title{Create a complete ggplot for opera's objects}
\usage{
ggopera(object, ...)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of ggOpera}

\item{...}{other arguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
Create a complete ggplot for opera's objects
}
