% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{iter.step}
\alias{iter.step}
\title{iterative step to calculate cox model derivatives}
\usage{
iter.step(Z, cov, time, cen, beta, theta, indvec)
}
\arguments{
\item{Z}{a matrix of staging term}

\item{cov}{a matrix of covariates}

\item{time}{a vector of survival outcome}

\item{cen}{a vector of censoring status}

\item{beta}{a vector of staging coefficients}

\item{theta}{a vector of covariates coefficients}

\item{indvec}{an indicator of risk set}
}
\value{
a list of derivatives
}
\description{
calculate the first and second order derivatives for IRLS of Cox model.
}
