% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{logPL}
\alias{logPL}
\title{calculate negative log likelihood of Cox model}
\usage{
logPL(Z, cov, cen, beta, theta)
}
\arguments{
\item{Z}{a matrix of staging term}

\item{cov}{a matrix of covariates}

\item{cen}{a vector of censoring status}

\item{beta}{a vector of staging coefficients}

\item{theta}{a vector of covariates coefficients}
}
\value{
negative log likelihood of Cox model
}
\description{
calculate negative log likelihood of Cox model
}
