% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lossConv}
\alias{lossConv}
\title{Error of a convex combination}
\usage{
lossConv(p, y, experts, awake = NULL, loss.type = "squareloss")
}
\arguments{
\item{p}{A non-negative vector that sum to
one. It is of length \code{N}, the number of experts.}

\item{y}{A vector that contains the
observations to be predicted.}

\item{experts}{A matrix containing the
experts forecasts. Each column corresponds to the predictions proposed by an
expert to predict \code{Y}. It has as many columns as there are experts.}

\item{awake}{A matrix specifying the
activation coefficients of the experts. Its entries lie in \code{[0,1]}.
Needed if some experts are specialists and do not always form and suggest
prediction.  If the expert number \code{k} at instance \code{t} does not
form any prediction of observation \code{Y_t}, we can put
\code{awake[t,k]=0} so that the mixture does not consider expert \code{k} in
the mixture to predict \code{Y_t}.}

\item{loss.type}{A string specifying
the loss function considered to evaluate the performance. It can be
"squareloss", "mae", "mape", or "pinballloss". See \code{\link{loss}} for
more details.}
}
\value{
The average errors suffered by the mixture. Note that if
\code{loss.type = "squareloss"}, the \code{rmse} is returned.  Note also
that the instance are weighted according to the number of activated experts.
}
\description{
The
function \code{lossConv} computes the loss suffered by the constant convex
mixture \code{p}.
}
\note{
The function \code{lossConv} is for instance used
to compute the best convex combination in hindsight.
}
\seealso{
\code{\link{loss}},
\code{\link{bestConvex}}
}
\keyword{~kwd1}
\keyword{~kwd2}

