% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(x, y, awake = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A vector containing the sequence of
predictions to be evaluated.}

\item{y}{A vector containing the sequence of
observations to be predicted.}

\item{awake}{A vector specifying the
activation coefficients of the predictions in \code{x}. Its entries lie in
\code{[0,1]}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Average absolute error of the sequence
}
\description{
The
function \code{mae} computes the average absolute error of a sequence.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{rmse}}, \code{\link{mape}}, \code{\link{loss}}
}
\keyword{~kwd1}
\keyword{~kwd2}

