% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A vector containing the sequence of
predictions to be evaluated.}

\item{y}{A vector containing the sequence of
observations to be predicted.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
The mean absolute percentage error of the sequence of
prediction \code{x}.
}
\description{
The
function \code{mape} computes the mean absolute percentage error of a
sequence of predictions. It is an equivalent way of running
\code{mean(loss(x,y,loss.type='mape'))} (see \code{\link{loss}}).
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{mae}}, \code{\link{rmse}}, \code{\link{loss}}
}
\keyword{~kwd1}
\keyword{~kwd2}

