% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{opera.solve}
\alias{opera.solve}
\title{solve opera odel}
\usage{
opera.solve(time, cen, Z, PO, cov, maxiter = 10, eps = 1e-04, GIC = 2)
}
\arguments{
\item{time}{a vector of follow-up time}

\item{cen}{a vector of censoring status}

\item{Z}{a matrix of staging variables}

\item{PO}{a vector of paritial order relationship staging variables}

\item{cov}{a matrix of covariates}

\item{maxiter}{maximum number of iterations in each approximation step}

\item{eps}{accuracy level}

\item{GIC}{the parameter of GIC law, the default value is 2, which is AIC law}
}
\value{
staging result of staging variables
}
\description{
Ordering poset element by recursive amalgamation.
Stratifying the staging variables with respect to the survival outcomes.
For computing efficiency, before inputting to functions, data should be sorted by follow-up time.
}
