% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pinballHour}
\alias{pinballHour}
\title{Follow the penalized leader aggregation rule for pinball loss}
\usage{
pinballHour(y, experts, lambda, href = 1, period = 1, w0 = NULL,
  tau = 0.5)
}
\arguments{
\item{y}{A vector containing the observations
to be predicted.}

\item{experts}{A matrix containing the
experts forecasts. Each column corresponds to the predictions proposed by an
expert to predict \code{Y}. It has as many columns as there are experts.}

\item{lambda}{A positive penalty
coefficient.}

\item{href}{A number in \code{[1,period]}
specifying the instant in the day when the aggregation rule can update its
weights.  It should lie in the interval \code{c(1,period)}.}

\item{period}{The number of instants in
each day.}

\item{w0}{A vector containing the prior
weights of the experts.}

\item{tau}{A number in \code{[0,1]}
describing the quantile to be predicted. Used only if \code{loss.type =
"pinballloss"}.}
}
\value{
\item{weights}{ A matrix of dimension \code{c(T,N)}, with
\code{T} the number of instances to be predicted and \code{N} the number of
experts.  Each row contains the convex combination to form the predictions.
} \item{prediction}{ A vector of length \code{T} that contains the quantiles
predictions outputted by the aggregation rule.  }
}
\description{
The
function \code{pinballHour} is a mixing aggregation rule for quantile
regression.  At each instance, it forms the mixture by performing a convex
minimisation. It chooses the mixture that minimizes among the past a
penalized criterion based on cumulated pinball loss.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{loss}}
}
\keyword{~kwd1}
\keyword{~kwd2}

