% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mixture.R
\name{plot.mixture}
\alias{plot.mixture}
\title{Plot an object of class mixture}
\usage{
\method{plot}{mixture}(
  x,
  pause = FALSE,
  col = NULL,
  alpha = 0.01,
  dynamic = FALSE,
  type = "all",
  max_experts = 50,
  col_by_weight = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  subset = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class mixture. If awake is provided (i.e., some experts are unactive), 
their residuals and cumulative losses are computed by using the predictions of the mixture.}

\item{pause}{if set to TRUE (default) displays the plots separately, otherwise on a single page}

\item{col}{the color to use to represent each experts, if set to NULL (default) use R\code{RColorBrewer::brewer.pal(...,"Spectral"}}

\item{alpha}{\code{numeric}. Smoothing parameter for contribution plot (parameter 'f' of function \code{\link[stats]{lowess}}).}

\item{dynamic}{\code{boolean}. If TRUE, graphs are generated with \code{rAmCharts}, else with base R.}

\item{type}{\code{char}.
\describe{
     \item{'all'}{ Display all the graphs ;}
     \item{'plot_weight', 'boxplot_weight', 'dyn_avg_loss', 'cumul_res', 'avg_loss', 'contrib'}{ Display the selected graph alone.}
}}

\item{max_experts}{\code{integer}. Maximum number of experts to be displayed (only the more influencial).}

\item{col_by_weight}{\code{boolean}. If TRUE (default), colors are ordered by weights of each expert, else by column}

\item{xlab}{\code{character}. Custom x-axis label (individual plot only)}

\item{ylab}{\code{character}. Custom y-axis label (individual plot only)}

\item{main}{\code{character}. Custom title (individual plot only)}

\item{subset}{\code{numeric}. Positive indices for subsetting data before plot.}

\item{...}{additional plotting parameters}
}
\value{
plots representing: plot of weights of each expert in function of time, boxplots of these weights,
cumulative loss \eqn{L_T=\sum_{t=1}^T l_{i,t}} of each expert in function of time, cumulative residuals \eqn{\sum_{t=1}^T (y_t-f_{i,t})} of each 
expert's forecast in function of time, average loss suffered by the experts and the contribution of each expert to the aggregation 
\eqn{p_{i,t}f_{i,t}} in function of time.
}
\description{
provides different diagnostic plots for an aggregation procedure.
}
\seealso{
See \code{\link{opera-package}} and opera-vignette for a brief example about how to use the package.
}
\author{
Pierre Gaillard <pierre.gaillard@inria.fr>

Yannig  Goude <yannig.goude@edf.fr>
}
