% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-oracle.R
\name{plot.oracle}
\alias{plot.oracle}
\title{Plot an aggregation procedure}
\usage{
\method{plot}{oracle}(x, sort = TRUE, col = NULL, dynamic = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{oracle}.}

\item{sort}{if set to TRUE (default), it sorts the experts by performance before the plots.}

\item{col}{colors}

\item{dynamic}{If TRUE, graphs are generated with \code{rAmCharts}, else with base R.}

\item{...}{additional arguments to function plot.}
}
\description{
oracle \code{plot}. It has one optional arguments.
}
