% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-oracle.R
\name{plt_oracle_convex}
\alias{plt_oracle_convex}
\title{Functions to render dynamic oracle graphs using rAmCharts}
\usage{
plt_oracle_convex(data, colors, round = 2)
}
\value{
a rAmCharts plot
}
\description{
Functions to render dynamic oracle graphs using rAmCharts
}
