% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ridgeCalib}
\alias{ridgeCalib}
\title{Ridge aggregation rule with automatic calibration of smoothing parameters}
\usage{
ridgeCalib(y, experts, gridlambda = 1, href = 1, period = 1, w0 = NULL,
  delay = 0, y.ETR = NULL, trace = F)
}
\arguments{
\item{y}{A vector containing the observations
to be predicted.}

\item{experts}{A matrix containing the
experts forecasts. Each column corresponds to the predictions proposed by an
expert to predict \code{Y}. It has as many columns as there are experts.}

\item{gridlambda}{A vector containing the initial possible values of the
smoothing parameter \code{lambda}.}

\item{href}{A number in \code{[1,period]}
specifying the instant in the day when the aggregation rule can update its
weights.  It should lie in the interval \code{c(1,period)}.}

\item{period}{The number of instants in
each day.}

\item{w0}{A vector containing the prior
weights of the experts.}

\item{delay}{A positive number that indicates the number of instants before
the mixture has access to the true observations in \code{y}.  If \code{delay
> 0}, \code{y.ETR} can not be \code{NULL}.}

\item{y.ETR}{A vector containing real time estimations of \code{y} to be
used at each instant \code{t} instead of \code{y_t} to predict instants
\code{t+1,...,t+delay}.}

\item{trace}{A boolean. If TRUE, the
evolution of the aggregation rule is displayed. Usefull if the code is too
long.}
}
\value{
\item{weights}{ A matrix of dimension \code{c(T,N)}, with
\code{T} the number of instances to be predicted and \code{N} the number of
experts.  Each row contains the convex combination to form the predictions.
} \item{prediction}{ A vector of length \code{T} that contains the quantiles
predictions outputted by the aggregation rule.  } \item{lambda}{ A vector of
length \code{T} containing the sequence of penalty coefficient chosen by the
aggregation rule.  } \item{grid}{ A vector containing the final grid of
potential penalty coefficient considered by the aggregation rule.  }
\item{loss}{ The error suffered by the aggregation rule determined by
\code{loss.type}.  If \code{loss.type = 'squareloss'}, the \link{rmse} is
computed.  } \item{gridloss}{ A vector of the same length as grid containing
the errors suffered by the \code{ridgeHour} aggregation rule if it
had picked the fixed learning rates in \code{grid}.  }
}
\description{
The function \code{ridgeCalib} performs ridge aggregation rule with
automatic calibration of the smoothing parameter \code{lambda} by performing
an optimization on a finite grid.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\keyword{~kwd1}
\keyword{~kwd2}

