% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(x, y, awake = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A vector containing the sequence of
predictions to be evaluated.}

\item{y}{A vector containing the sequence of
observations to be predicted.}

\item{awake}{A vector specifying the
activation coefficients of the predictions in \code{x}. Its entries lie in
\code{[0,1]}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
The root mean square error of the sequence of prediction
\code{x}.
}
\description{
The
function \code{rmse} computes the root mean square error of a sequence of
predictions. It is an equivalent way of running
\code{sqrt(mean(loss(x,y,loss.type='squareloss')))} (see
\code{\link{loss}}).
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{mape}}, \code{\link{mae}}, \code{\link{loss}}
}
\keyword{~kwd1}
\keyword{~kwd2}

