% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rowSums}
\alias{rowSums}
\title{calculate sum of rows}
\usage{
rowSums(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{a matrix to calculate rowsums}
}
\value{
a vector of row sums
}
\description{
calculate sum of rows
}
