% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-mixture.R
\name{mixtureUI}
\alias{mixtureUI}
\alias{mixtureServer}
\title{User Interface for exploring mixture result}
\usage{
mixtureUI(id, plotly = FALSE)

mixtureServer(input, output, session, mixture, type, ncol = 2, plotly)
}
\arguments{
\item{id}{Module's Id}

\item{plotly}{Logical, add a dependance to plotly ?}

\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{mixture}{a reactive function containing an object of class mixture}

\item{type}{which graphic to plot, a value between 1 and 6, can be a reactive function}

\item{ncol}{number of column for displaying plots}

\item{plotly}{logical, use plotly to plot output, can be a reactive function}
}
\value{
mixtureUI : a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
User Interface for exploring mixture result

Server for exploring mixture result
}
\examples{
\dontrun{
library("shiny")
runApp(system.file("modules/module-mixture.R", package = "opera"))
}
}
