% Generated by roxygen2 (4.0.1): do not edit by hand
\name{truncate1}
\alias{truncate1}
\title{truncate1 a real number}
\usage{
truncate1(x)
}
\arguments{
\item{x}{A number to be truncate1d}
}
\value{
The truncated value of \code{x}
}
\description{
The function \code{truncate1} is used to avoid instability in \code{R} due to
very smal or very big numbers.  It truncate1s numbers so that they lie
between \code{exp(-700)} and \code{exp(700)}.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\keyword{~kwd1}
\keyword{~kwd2}

