% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loss}
\alias{loss}
\title{Error suffered by a sequence of prediction}
\usage{
loss(x, y, loss.type = "squareloss", tau = 0.1)
}
\arguments{
\item{x}{A vector of length \code{T}
containing the sequence of prediction to be evaluated.}

\item{y}{A vector of length \code{T} that
contains the observations to be predicted.}

\item{loss.type}{A string specifying
the loss function considered to evaluate the performance. It can be
"squareloss", "mae", "mape", or "pinballloss". See \code{\link{loss}} for
more details.}

\item{tau}{A number in \code{[0,1]}
describing the quantile to be predicted. Used only if \code{loss.type =
"pinballloss"}.}
}
\value{
A vector of length \code{T} containing the sequence of
instantaneous losses suffered by the prediction \code{x}.
}
\description{
The
function \code{loss} computes the sequence of instantaneous losses suffered
by the predictions in \code{x} to predict the observation in \code{y}.
}
\author{
Pierre Gaillard <pierre-p.gaillard@edf.fr>
}
\seealso{
\code{\link{mape}}, \code{\link{rmse}}, \code{\link{lossConv}}
}
\keyword{~kwd1}
\keyword{~kwd2}

