% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-oracle.R
\name{plt_oracle_convex}
\alias{plt_oracle_convex}
\title{Functions to render dynamic oracle graphs using rAmCharts}
\usage{
plt_oracle_convex(data, colors, round = 2)
}
\arguments{
\item{data}{\code{named vector}. Vector of values to be displayed.}

\item{colors}{\code{character}. Colors to  be used.}

\item{round}{\code{integer} (2). Precision of the values in the tooltips..}
}
\value{
a rAmCharts plot
}
\description{
Functions to render dynamic oracle graphs using rAmCharts
}
